Gui: Move color traits to Utilities.h
